+++ package.json
@@ -6,9 +6,13 @@
     "react": "^16.13.1",
     "react-dom": "^16.12.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1"
+    "react-scripts": "3.4.1",
+    "typescript": "^4.0.3"
   },
   "devDependencies": {
+    "@types/react": "^16.9.50",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-router-dom": "^5.1.5",
     "react-scripts": "3.3.0"
   },
   "scripts": {
