--- src/App.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const App = () => (
-  <div className="App" />
-)
-
-export default App
