+++ src/App.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const App = () => (
+  <div className="App" />
+);
+
+export default App;
